+++ enclone_main/src/lib.rs
@@ -9,6 +9,7 @@ pub mod inconsistent;
 pub mod main_enclone;
 pub mod merge_onesies;
 pub mod populate_features;
+pub mod sec_mem;
 pub mod setup;
 pub mod split_orbits;
 pub mod vars;
