+++ enclone_main/src/main_enclone.rs
@@ -11,6 +11,7 @@ use crate::flag_defective::*;
 use crate::inconsistent::*;
 use crate::merge_onesies::*;
 use crate::populate_features::*;
+use crate::sec_mem::*;
 use crate::setup::*;
 use crate::split_orbits::*;
 use crate::vars::*;
@@ -44,7 +45,6 @@ use std::{
     env,
     fs::File,
     io::{BufRead, BufReader, BufWriter, Write},
-    process::Command,
     thread, time,
     time::Instant,
 };
@@ -270,86 +270,10 @@ pub fn main_enclone(args: &Vec<String>) {
 
     ctl.gen_opt.species = species(&refdata);
 
-    // Test for okness of sec/mem args.
+    // Process for sec (secreted) or mem (membrane) if specified.
 
-    let mut vars = ctl.parseable_opt.pcols.clone();
-    vars.append(&mut ctl.clono_print_opt.lvars.clone());
-    unique_sort(&mut vars);
-    ctl.gen_opt.using_secmem =
-        bin_member(&vars, &"sec".to_string()) || bin_member(&vars, &"mem".to_string());
-    if !ctl.gen_opt.using_secmem
-        && ctl.parseable_opt.pout.len() > 0
-        && ctl.parseable_opt.pcols.len() == 0
-    {
-        if ctl.gen_opt.species == "human" || ctl.gen_opt.species == "mouse" {
-            if is_bcr {
-                let mut have_bam = true;
-                for g in ctl.origin_info.gex_path.iter() {
-                    if g.len() == 0 {
-                        have_bam = false;
-                        break;
-                    }
-                    let bam = format!("{}/possorted_genome_bam.bam", g);
-                    if !path_exists(&bam) {
-                        have_bam = false;
-                        break;
-                    }
-                }
-                if have_bam {
-                    let o = Command::new("samtools")
-                        .arg("--help")
-                        .output()
-                        .expect("failed to execute samtools");
-                    let status = o.status.code().unwrap();
-                    if status == 0 {
-                        ctl.gen_opt.using_secmem = true;
-                    }
-                }
-            }
-        }
-    }
-    if bin_member(&vars, &"sec".to_string()) || bin_member(&vars, &"mem".to_string()) {
-        if ctl.gen_opt.species != "human" && ctl.gen_opt.species != "mouse" {
-            eprintln!("\nThe lvars sec and mem can only be used for data from human and mouse.\n");
-            std::process::exit(1);
-        }
-        if !is_bcr {
-            eprintln!("\nThe lvars sec and mem do not make sense for TCR data.\n");
-            std::process::exit(1);
-        }
-        for g in ctl.origin_info.gex_path.iter() {
-            if g.len() == 0 {
-                eprintln!("\nThe lvars sec and mem can only be used if GEX data are provided.\n");
-                std::process::exit(1);
-            }
-            let bam = format!("{}/possorted_genome_bam.bam", g);
-            if !path_exists(&bam) {
-                eprintln!(
-                    "\nThe lvars sec and mem can only be used if the file\n\
-                    pos_sorted_genome_bam.bam is provided.  We did not see it at this path\n\
-                    {}.",
-                    g
-                );
-                std::process::exit(1);
-            }
-        }
-        let o = Command::new("samtools")
-            .arg("--help")
-            .output()
-            .expect("failed to execute samtools");
-        let status = o.status.code().unwrap();
-        if status != 0 {
-            eprintln!(
-                "\nThe lvars sec and mem can only be used if the samtools\n\
-                executable is in your path.\n"
-            );
-            std::process::exit(1);
-        }
-    }
+    test_sec_mem(&mut ctl);
     ctl.perf_stats(&tr, "building reference and other things");
-
-    // If sec (secreted) or mem (membrane) lvars have been specified, gather those data.
-
     if ctl.gen_opt.using_secmem {
         fetch_secmem(&mut ctl);
     }
